feature/3402 adjust and replace floodingmap wuppertal with playground version adjust paths and deploymentconfig
