+++ envirometrics/wuppertal/floodingmap/README.md
@@ -1,3 +1,12 @@
 # TopicMap Hochwasser in Wuppertal
 
-<img width="1488" alt="Screenshot 2023-08-23 at 16 50 36" src="https://github.com/topicmaps-wuppertal/hochwasser/assets/837211/736cc19c-6504-4fff-81d5-cc8cbaab2c87">
+Flooding Map (Hochwassergefahrenkare HGK) with integrated 3D component extending the EnviroMetricsMap (2D)
+
+
+
+Changes vs old Envirometrics:
+
+to enable Cesium CustomViewer lib added:
+- Redux-Store,
+- HashRouter
+- Providers
