+++ envirometrics/wuppertal/floodingmap/public/data/glb/map_pointer.glb
