+++ envirometrics/wuppertal/floodingmap/src/config/cesium/assets.config.ts
