--- envirometrics/wuppertal/floodingmap/src/index.css
@@ -1,30 +1,0 @@
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
-
-.h4,
-h4 {
-  font-size: 18px !important;
-}
-
-.badge-primary {
-  color: #fff;
-  background-color: #337ab7 !important;
-}
-.leaflet-container .leaflet-control-attribution {
-  margin-right: 11px !important;
-}
-
-svg path.leaflet-interactive {
-  outline: none;
-}
