+++ envirometrics/wuppertal/floodingmap/src/main.tsx
@@ -1,25 +1,76 @@
-import { StrictMode } from "react";
-import * as ReactDOM from "react-dom/client";
-import App from "./App.jsx";
+import { createRoot } from "react-dom/client";
+import { createHashRouter, RouterProvider } from "react-router-dom";
 
-const originalWarn = console.warn.bind(console);
-const originalError = console.error.bind(console);
-console.warn = (message, ...args) => {
-  if (!message.includes("ReactDOM.render is no longer supported in React 18")) {
-    originalWarn(message, ...args);
-  }
-};
-console.error = (message, ...args) => {
-  if (!message.includes("ReactDOM.render is no longer supported in React 18")) {
-    originalError(message, ...args);
-  }
-};
+import { Provider } from "react-redux";
+import { persistStore } from "redux-persist";
+import { PersistGate } from "redux-persist/integration/react";
 
-const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
+import { MappingConstants } from "react-cismap";
+import { TopicMapContextProvider } from "react-cismap/contexts/TopicMapContextProvider";
+import { CrossTabCommunicationContextProvider } from "react-cismap/contexts/CrossTabCommunicationContextProvider";
+
+import { GazDataProvider, SelectionProvider } from "@carma-apps/portals";
+import { TweakpaneProvider } from "@carma-commons/debug";
+import { suppressReactCismapErrors } from "@carma-commons/utils";
+import {
+  CesiumContextProvider,
+  setupCesiumEnvironment,
+} from "@carma-mapping/cesium-engine";
+
+import App from "./App";
+import store from "./store";
+import { prefix, sourcesConfig } from "./config/gazData";
+import { SYNC_TOKEN } from "./config/app.config";
+import { CESIUM_CONFIG } from "./config/cesium/cesium.config";
+
+suppressReactCismapErrors();
+setupCesiumEnvironment();
+
+const persistor = persistStore(store);
+
+const enableSync = true;
+
+const syncedApp = (
+  <CrossTabCommunicationContextProvider role="sync" token={SYNC_TOKEN}>
+    <App sync={true} />
+  </CrossTabCommunicationContextProvider>
 );
+
+const appWithContext = (
+  <GazDataProvider sourcesConfig={sourcesConfig} prefix={prefix}>
+    <SelectionProvider>
+      <TopicMapContextProvider
+        appKey={"Hochwasserkarte.Story.Wuppertal"}
+        referenceSystem={MappingConstants.crs3857}
+        referenceSystemDefinition={MappingConstants.proj4crs3857def}
+        // baseLayerConf={wuppertalConfig.overridingBaseLayerConf}
+        infoBoxPixelWidth={370}
+      >
+        <CesiumContextProvider
+          providerConfig={CESIUM_CONFIG.providerConfig}
+          tilesetConfigs={CESIUM_CONFIG.tilesetConfigs}
+        >
+          {enableSync ? syncedApp : <App />}
+        </CesiumContextProvider>
+      </TopicMapContextProvider>
+    </SelectionProvider>
+  </GazDataProvider>
+);
+
+const router = createHashRouter([
+  {
+    path: "/",
+    element: appWithContext,
+  },
+]);
+const root = createRoot(document.getElementById("root") as HTMLElement);
+
 root.render(
-  <StrictMode>
-    <App />
-  </StrictMode>
+  <Provider store={store}>
+    <TweakpaneProvider>
+      <PersistGate loading={null} persistor={persistor}>
+        <RouterProvider router={router} />
+      </PersistGate>
+    </TweakpaneProvider>
+  </Provider>
 );
