+++ envirometrics/wuppertal/floodingmap/src/utils/cesiumHandlers.ts
