+++ envirometrics/wuppertal/floodingmap/src/utils/marker.ts
