+++ envirometrics/wuppertal/floodingmap/src/utils/scene.ts
