+++ envirometrics/wuppertal/floodingmap/src/version.json
@@ -1,3 +1,3 @@
 {
-  "version": "1.0.2"
+  "version": "2.0.0"
 }
