+++ envirometrics/wuppertal/floodingmap/tailwind.config.cjs
@@ -6,7 +6,7 @@ module.exports = {
     path.join(__dirname, "./src/**/*.{js,ts,cjs,mjs,tjs,jsx,tsx}"),
     path.join(
       __dirname,
-      "../../libraries/mapping/carma-map-layers/src/**/*.{js,ts,jsx,tsx}",
+      "../../../libraries/mapping/carma-map-layers/src/**/*.{js,ts,jsx,tsx}",
     ),
   ],
   theme: {
