--- playgrounds/floodingmap/README.md
@@ -1,18 +1,0 @@
-# TopicMap Hochwasser in Wuppertal
-
-Playground for integrating a 3D component with EnviroMetricsMap (2D)
-see also CarmaMap Playground
-
-<img width="1488" alt="Screenshot 2023-08-23 at 16 50 36" src="https://github.com/topicmaps-wuppertal/hochwasser/assets/837211/736cc19c-6504-4fff-81d5-cc8cbaab2c87">
-
-Changes vs Envirometrics:
-
-add viteStaticCopy to vite.config.plugins
-change vite.config.ts to vite.config.mts
-
-to enable Cesium CustomViewer lib added:
-- Redux-Store,
-- HashRouter
-- Providers
-
-add 3d marker asset to public/data
