--- playgrounds/floodingmap/project.json
@@ -1,67 +1,0 @@
-{
-  "name": "floodingmap-playground",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "playgrounds/floodingmap/src",
-  "projectType": "application",
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/vite:build",
-      "outputs": ["{options.outputPath}"],
-      "defaultConfiguration": "production",
-      "options": {
-        "outputPath": "dist/playgrounds/floodingmap"
-      },
-      "configurations": {
-        "development": {
-          "mode": "development"
-        },
-        "production": {
-          "mode": "production"
-        }
-      }
-    },
-    "serve": {
-      "executor": "@nx/vite:dev-server",
-      "defaultConfiguration": "development",
-      "options": {
-        "buildTarget": "floodingmap-playground:build"
-      },
-      "configurations": {
-        "development": {
-          "buildTarget": "floodingmap-playground:build:development",
-          "hmr": true
-        },
-        "production": {
-          "buildTarget": "floodingmap-playground:build:production",
-          "hmr": false
-        }
-      }
-    },
-    "preview": {
-      "executor": "@nx/vite:preview-server",
-      "defaultConfiguration": "development",
-      "options": {
-        "buildTarget": "floodingmap-playground:build"
-      },
-      "configurations": {
-        "development": {
-          "buildTarget": "floodingmap-playground:build:development"
-        },
-        "production": {
-          "buildTarget": "floodingmap-playground:build:production"
-        }
-      }
-    },
-    "test": {
-      "executor": "@nx/vite:test",
-      "outputs": ["{options.reportsDirectory}"],
-      "options": {
-        "reportsDirectory": "../../coverage/playgrounds/floodingmap"
-      }
-    },
-    "lint": {
-      "executor": "@nx/eslint:lint"
-    }
-  }
-}
