--- playgrounds/floodingmap/src/components/notification.css
@@ -1,3 +1,0 @@
-.ant-notification-stack > .ant-notification-notice-wrapper {
-  position: relative;
-}
