--- playgrounds/floodingmap/src/styles.css
@@ -1,24 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-/* override tailwind base img display: image reset for proper infobox layout */
-.leaflet-control * img {
-  display: initial;
-}
-
-/* intermediate workaround for antd tooltip arrow to not show until bootstrap style interference is resolvedß */
-.ant-tooltip-arrow {
-  display: none !important;
-}
-
-/* quite specific workaround for selecting the UI should have classname to select from not the title */
-.isMode3d a[title="Stadtplan"],
-.isMode3d a[title="Top. Karte"] {
-  opacity: 0.5;
-  pointer-events: none;
-}
-
-.modal-footer img[alt="Logo DigiTal Zwilling"] {
-  max-height: 60px;
-}
