--- playgrounds/floodingmap/src/version.json
@@ -1,3 +1,0 @@
-{
-  "version": "2.0.0"
-}
