--- playgrounds/floodingmap/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "../../tsconfig.legacy.base.json",
-  "compilerOptions": {},
-  "files": [],
-  "references": []
-}
